// packages/layout/src/insets/Types.ts
export type ContainerInsets = {
  readonly bottom: number;
  readonly left: number;
  readonly right: number;
  readonly top: number;
}

export type ContainerOffsets = {
  readonly offsetBottom: number;
  readonly offsetLeft: number;
  readonly offsetRight: number;
  readonly offsetTop: number;
}

export type ElementRect = {
  readonly bottom: number;
  readonly height: number;
  readonly left: number;
  readonly right: number;
  readonly top: number;
  readonly width: number;
  readonly x: number;
  readonly y: number;
}